% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/AndelerTableContainer.R
\name{AndelerTableContainer}
\alias{AndelerTableContainer}
\title{Make html table container for DT used in Shiny}
\usage{
AndelerTableContainer(groupText = "Gruppe", deptName = "ShusAvd")
}
\arguments{
\item{groupText}{String General name of grouping}

\item{deptName}{String Name of department where data belongs}
}
\description{
Yes, make it!
}

