% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeHighcharts.R
\name{AndelTidHighchart}
\alias{AndelTidHighchart}
\title{Provider of Highchart object}
\usage{
AndelTidHighchart(Aartxt, AndelHoved, NAarHendHoved, NHovedRes, AndelRest,
  NAarHendRest, NSmlRes, Tittel, utvalgTxt, shtxt, smltxt, fargeHoved,
  fargeRest)
}
\value{
h1 Highchart object
}
\description{
Provide Highchart object for AndelTid
}

