% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{dataTilOffVisning}
\alias{dataTilOffVisning}
\title{Tilrettelegge data for offentlig visning.}
\usage{
dataTilOffVisning(
  RegData = RegData,
  valgtVar,
  datoFra = "2014-01-01",
  aar = 0,
  ResPort = 1,
  filUt = "dummy",
  lagreFil = 1
)
}
\arguments{
\item{RegData}{- data}

\item{valgtVar}{- 'KomplSvelging3mnd', 'KomplStemme3mnd', 'Komplinfek', 'NDIendr12mnd35pst}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{filUt}{tilnavn for utdatatabell (fjern?)}

\item{lagreFil}{automatisk lagre fila som genereres? 0-nei, 1-ja (standard)}
}
\value{
Datafil til Resultatportalen
}
\description{
Tilrettelegge data for offentlig visning.
}
