% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{dataTilResPort}
\alias{dataTilResPort}
\title{Generere data til Resultatportalen}
\usage{
dataTilResPort(
  RegData = RegData,
  valgtVar,
  datoFra = "2014-01-01",
  aar = 0,
  myelopati = 99,
  fremBak = 0,
  filUt = "dummy"
)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- beinsmLavPre, peropKompDura, sympVarighUtstr, NDIendr12mnd35pst}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{myelopati}{Operert for myelopati. 0-nei, 1-ja, 99-alle (standard)}

\item{fremBak}{Fremre eller bakre tilgang. 0-ikke valgt, 1-fremre, 2-bakre}

\item{filUt}{tilnavn for utdatatabell (fjern?)}
}
\value{
Datafil til Resultatportalen
}
\description{
Generere data til Resultatportalen
}
