% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeHighcharts.R
\name{EmptyHighchart}
\alias{EmptyHighchart}
\title{Provider of Highchart object}
\usage{
EmptyHighchart(Tittel, utvalg = "", infoText = "Tomt...")
}
\arguments{
\item{Tittel}{String Title to show in empty chart}

\item{infoText}{String Text to shown in empty chart}
}
\value{
h1 Highchart object
}
\description{
Provide Highchart object when no data
}

