% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndelerGrVar.R
\name{FigAndelerGrVar}
\alias{FigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
FigAndelerGrVar(RegData, valgtVar, datoFra = "2013-01-01",
  datoTil = "3000-12-31", enhetsUtvalg = 0, minald = 0, maxald = 130,
  erMann = "", hentData = 0, tittel = 1, reshID, outfile = "")
}
\arguments{
\item{minald}{- alder}
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt grupperingsvariabel,
f.eks. sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
\details{
Detajer:
}

