% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndelTid.R
\name{FigAndelTid}
\alias{FigAndelTid}
\title{Tidstrend (år) av andel for en gitt variabel.}
\usage{
FigAndelTid(RegData, valgtVar, datoFra = "2013-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  tittel = 1, reshID, outfile = "", enhetsUtvalg = 1, hentData = 1)
}
\arguments{
\item{RegData}{- data}
}
\description{
Funksjon som genererer en figur med andeler av en variabel for hvert år.
}
\details{
Detaljer:
}

