% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigGjsnTid.R
\name{FigGjsnTid}
\alias{FigGjsnTid}
\title{Tidsutvikling av gjennomsnitt/median for en gitt variabel}
\usage{
FigGjsnTid(RegData, outfile, valgtVar, erMann = "", minald = 0,
  maxald = 130, tittel = 1, datoFra = "2007-01-01",
  datoTil = "3000-01-01", valgtMaal = "", enhetsUtvalg = 1,
  hentData = 0, reshID)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{outfile}{- navn på fil figuren skrives ned til}

\item{valgtVar}{- Variabelen det skal vises resultat for. Må velges: Alder, BMI, ...}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{datoFra}{- Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{- Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{valgtMaal}{- 'Med' gir median, alt annet gir gjennomsnitt}

\item{enhetsUtvalg}{- 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{hentData}{- Settes til 1 (standard) om det skal spørres etter data fra funksjonen.
Settes til en annen verdi om data leveres til funksjonen gjennom 'RegData',
eksempelvis ved bruk av eksempeldatasettet eller ved kall fra andre
funksjoner der data allerede er tilgjengelig.}

\item{reshID}{- avdelingsid (reshID) for egen avdeling,}
}
\description{
Gjennomsnitt/median per år med konfidensintervall for valgt variabel.
Hvis man har valgt å sammenlikne, vises  konfidensintervall for resten av landet (evt. annen
sammenlikningsgruppe) i bakgrunnen.
}
\details{
Smerte: Skalaen går fra 0 til 10, dvs. at differansen ligger mellom -10 og 10..
EQ5D: Skala fra -0.594 tl 1, jo høyere jo bedre.
Oswestry: Skala fra 0 til 100, hvor lavest er friskest
}

