% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigGjsnMed.R
\name{FigMeanMed}
\alias{FigMeanMed}
\title{Søylediagram som viser sentralmål (gj.sn./median) for hvert sykehus}
\usage{
FigMeanMed(RegData, valgtVar, valgtMaal = "Gjsn", datoFra = "2012-04-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = "", reshID,
  outfile = "", hentData = 0)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{- Variabelen det skal vises resultat for. Må velges: Alder, BMI, ...}

\item{valgtMaal}{- 'Med' gir median, alt annet gir gjennomsnitt}

\item{datoFra}{- Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{- Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{reshID}{- avdelingsid (reshID) for egen avdeling,}

\item{outfile}{- navn på fil figuren skrives ned til}

\item{hentData}{- Settes til 1 (standard) om det skal spørres etter data fra funksjonen.
Settes til en annen verdi om data leveres til funksjonen gjennom 'RegData',
eksempelvis ved bruk av eksempeldatasettet eller ved kall fra andre
funksjoner der data allerede er tilgjengelig.}
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}

