% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeHighcharts.R
\name{GjsnGrVarHighchart}
\alias{GjsnGrVarHighchart}
\title{Provider of Highchart object}
\usage{
GjsnGrVarHighchart(Midt, N, Ngrtxt, tittel, utvalgTxt, GrNavnSort, xaksetxt,
  KIHele, KIopp, KIned, MidtHele, farger, deltittel)
}
\value{
h1 Highchart object
}
\description{
Provide Highchart object for GjsnGrVar
}

