% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{lageTulleData}
\alias{lageTulleData}
\title{Lage tulledata (simulerte data) for åpen publisering}
\usage{
lageTulleData(RegData, varBort = NA, antSh = 26, antObs = 20000)
}
\description{
Lage tulledata (simulerte data) for åpen publisering
}
