% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndeler.R
\name{NakkeFigAndeler}
\alias{NakkeFigAndeler}
\title{Søylediagram, horisontalt eller vertikalt, som viser andeler av valgt variabel.}
\usage{
NakkeFigAndeler(
  RegData = 0,
  valgtVar = "Alder",
  erMann = "",
  datoFra = "2012-01-01",
  datoTil = "3000-12-31",
  inngrep = 99,
  minald = 0,
  maxald = 110,
  myelopati = 99,
  fremBak = 0,
  outfile = "",
  hentData = 0,
  preprosess = 0,
  reshID = 0,
  enhetsUtvalg = 0,
  lagFig = 1,
  ...
)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Variabelen det skal vises resultat for.
            Alder: Aldersfordeling
            AntallNivaaOpr: Antall nivå operert
            Antibiotika: Er det gitt antibiotikaprofylakse?
            Arbeidstaus12mnd: Arbeidsstatus 12 mnd. etter operasjon
            Arbeidstaus3mnd: Arbeidsstatus 3 mnd. etter operasjon
            ArbeidstausPreOp: Arbeidsstatus før operasjon
            ASAgrad: ASA-grad
            BMI: Pasientenes BMI (Body Mass Index)
            EqAngstPreOp: Helsetilstand, Angst
            ErstatningPreOp: Søkt erstatning?
            FornoydBeh12mnd: Fornøydhet med behandlinga på sykehuset, 12 mnd
            FornoydBeh3mnd: Fornøydhet med behandlinga på sykehuset, 3 mnd
            OperasjonsKategori: Hastegrad
            LiggeDognPostop: Antall liggedøgn postoperativt
            LiggeDognTotalt: Totalt antall liggedøgn
            Morsmal: Morsmål
            NytteOpr12mnd: Nytte av operasjon, 12 mnd
            NytteOpr3mnd: Nytte av operasjon, 3 mnd
            OprIndikPareseGrad: Paresegrad før operasjon
            Roker: Røyker pasienten?
            Saardren: Har pasienten fått sårdren?
            SivilStatus: Sivilstatus
            SmertestillBrukPreOp: Hyppighet av smertestillende før operasjonen
            Snuser: Snuser pasienten?
            SymptVarighetArmer: Varighet av utstrålende armsmerter
            SymptVarighetNakkeHode: Varighet av nakke-/hodesmerter
            TidlOpr: Er pasienten tidligere operert?
            TidlOprAntall: Antall tidligere operasjoner
            UforetrygdPreOp: Søkt uføretrygd?
            Utdanning: Utdanningsnivå

Detajer...:}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{inngrep}{Inngrepstype     0-'Ikke klassifiserbar operasjon', 1-'Fremre diketomi for prolaps',
2-'Bakre dekompresjon', 3-'Fremre dekompresjon sp st.uten prolaps', 4-'Bakre fusjon',
5-'Korporektomi', 6-'Andre inngrep', 99-alle (standard)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{myelopati}{Operert for myelopati. 0-nei, 1-ja, 99-alle (standard)}

\item{fremBak}{Fremre eller bakre tilgang. 0-ikke valgt, 1-fremre, 2-bakre}

\item{outfile}{Navn på fil figuren skrives ned til}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}
}
\value{
En figur med søylediagram (fordeling) av ønsket variabel
}
\description{
Søylediagrammet viser fordelinga til den valgte variabelen. Søylene er horisontale eller vertikale
avhengig av hvor stor plass kategorinavnet til søyla tar.
}
