% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndelTid.R
\name{NakkeFigAndelTid}
\alias{NakkeFigAndelTid}
\title{Tidstrend (år) av andel for en gitt variabel.}
\usage{
NakkeFigAndelTid(RegData, valgtVar = "Alder", datoFra = "2013-01-01",
  datoTil = "3000-12-31", tidsenhet = "Aar", minald = 0,
  maxald = 130, erMann = "", myelopati = 99, fremBak = 0,
  reshID = 0, outfile = "", enhetsUtvalg = 0, preprosess = TRUE,
  hentData = 0)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Variabelen det skal vises resultat for.
Alder: Aldersfordeling
AndreRelSykdommer: Andre sykdommer
Antibiotika: Fått antibiotika
Arbeidstaus12mnd: Mottar sykepenger, 12 mnd etter operasjon?
Arbeidstaus3mnd: Mottar sykepenger, 3 mnd etter operasjon?
ArbeidstausPreOp: Mottar sykepenger, preoperativt?
ASAgrad: ASA-grad > II
BMI: Pasienter med fedme
EnhverKompl3mnd: Alle komplikasjoner
ErstatningPreOp: Søkt/planlegger å søke erstatning
FornoydBeh12mnd: Fornøyde pasienter, 12 mnd.
FornoydBeh3mnd: Fornøyde pasienter, 3 mnd.
KomplinfekDyp3mnd: Pasientrapportert dyp infeksjon, 3 mnd.
KomplinfekOverfl3mnd: Overfladisk infeksjon, 3 mnd.
KomplStemme3mnd: Stemmevansker, 3 mnd.
KomplSvelging3mnd: Svelgvansker, 3 mnd.
Misfor12mnd: Misfornøyde pasienter, 12 mnd.
Misfor3mnd: Misfornøyde pasienter, 3 mnd.
NytteOpr12mnd: Klart bedre, 12 mnd.
NytteOpr3mnd: Klart bedre, 3 mnd.
OprIndikMyelopati: Operasjonsårsak, Myelopati
OprIndikSmerter: Operasjonsårsak, Smerter
PerOpEnhverKompl: Komplikasjoner ved operasjon
Roker: Røykere
Saardren: Andel som får sårdren
SmertestillPreOp: Bruker smertestillende, preop.
SymptVarighetNakkeHode: Varighet av hode-/nakkesmerter over 1 år
SymptVarighetArmer: Varighet av armsmerter, minst 1 år  #' SymptVarighetSmerterUker
UforetrygdPreOp: Søkt eller planlegger å søke uføretrygd?
Utdanning: Andel høyskole-/universitetsutdannede
Verre12mnd: Klart verre, 12 mnd.
Verre3mnd. Klart verre, 3 mnd.}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{tidsenhet}{Oppløsning på tidsaksen. Verdier: 'Aar' (standard), 'Halvaar', 'Kvartal','Mnd'}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{myelopati}{Operert for myelopati. 0-nei, 1-ja, 99-alle (standard)}

\item{fremBak}{Fremre eller bakre tilgang. 0-ikke valgt, 1-fremre, 2-bakre}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{outfile}{Navn på fil figuren skrives ned til}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}
}
\value{
Figur med ...
}
\description{
Denne funksjonen lager et linjediagram som viser utvikling over tid  for andeler av valgt variabel,
filtrert på de utvalg som er gjort.
}
\details{
Detaljer
}
