% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigGjsnGrVar.R
\name{NakkeFigGjsnGrVar}
\alias{NakkeFigGjsnGrVar}
\title{Søylediagram som viser sentralmål (gj.sn./median) for hvert sykehus}
\usage{
NakkeFigGjsnGrVar(RegData, valgtVar, valgtMaal = "Gjsn",
  datoFra = "2012-01-01", datoTil = "2050-12-31", myelopati = 99,
  fremBak = 0, Ngrense = 10, medKI = 1, minald = 0, maxald = 130,
  erMann = "", reshID = 0, outfile = "", hentData = 0,
  preprosess = TRUE)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Variabelen det skal vises resultat for.
Alder: alder (år)
EMSscorePreOp: EMS hos Myelopatipasienter før
EMSendr12mnd: Forbedring av EMS hos myelopati-pasienter, 12 mnd.
EMSendr3mnd: Forbedring av EMS hos myelopati-pasienter, 3 mnd.
EQ5Dendr12mnd: Forbedring av EQ5D, 12 mnd.
EQ5Dendr3mnd: Forbedring av EQ5D, 3 mnd.
KnivtidTotalMin: total knivtid
NDIscorePreOp: NDI før operasjon
NDIendr3mnd:
LiggeDognPostop: liggetid etter operasjon
LiggeDognTotalt: antall liggedøgn, totalt
NRSsmerteArmPreOp: NSR, arm før operasjon
NRSsmerteNakkePreOp: NSR, nakke før operasjon}

\item{valgtMaal}{Sentralmål 'Med' gir median, alt annet gir gjennomsnitt}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{myelopati}{Operert for myelopati. 0-nei, 1-ja, 99-alle (standard)}

\item{fremBak}{Fremre eller bakre tilgang. 0-ikke valgt, 1-fremre, 2-bakre}

\item{Ngrense}{nedre grense for antall observasjoner for at et resultat (ei søyle) skal vises}

\item{medKI}{vise konfidensintervall? 0-nei, 1-ja}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{outfile}{Navn på fil figuren skrives ned til}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}
}
\value{
Figur med...
}
\description{
Funksjon som genererer en figur med gjennomsnitt/median
for hvert sykehus og kan ta inn ulike numeriske variable.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}
