% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHentData.R
\name{NakkeHentRegData}
\alias{NakkeHentRegData}
\title{Henter Nakke-tabeller og kobler sammen}
\usage{
NakkeHentRegData(
  datoFra = "2013-01-01",
  datoTil = Sys.Date(),
  medOppf = 1,
  ...
)
}
\arguments{
\item{datoFra}{fra og med operasjonsdato, format: 'yyyy-mm-dd'}

\item{datoTil}{til og med operasjonsdato, format: 'yyyy-mm-dd'}

\item{medOppf}{ha med oppfølgingsskjema? 0-nei, 1-ja}

\item{medPROM:}{koble på RAND og TSS2-variabler}

\item{alleData}{1- alle variabler med, 0 - utvalgte variabler med}
}
\value{
RegData data frame
}
\description{
Henter Nakke-tabeller og kobler sammen
}
