% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkePreprosesser.R
\name{NakkePreprosess}
\alias{NakkePreprosess}
\title{Preprosesser data fra Degenerativ Nakke}
\usage{
NakkePreprosess(RegData = RegData)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}
}
\value{
RegData En dataramme med det preprosesserte datasettet
}
\description{
Denne funksjonen definerer og formaterer variabler
}

