% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeRegDataSQL.R
\name{NakkeRegDataSQL}
\alias{NakkeRegDataSQL}
\title{Henter data registrert for Degenerativ Nakke}
\usage{
NakkeRegDataSQL(
  datoFra = "2012-01-01",
  datoTil = Sys.Date(),
  medProm = 1,
  alleVar = 0
)
}
\arguments{
\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{medProm}{Ha med prom-skjema 0-nei, 1-ja (standard)}

\item{alleVar}{1 - alle variabler med,
0 -bare variabler som er i bruk på Rapporteket.}
}
\value{
Henter dataramma RegData for Degenerativ Nakke
}
\description{
Henter data for Degenerativ Nakke fra "staging"-database.
Kan benytte datoFra og datoFra som input.
}
