% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakkeStandard.R
\name{nakkeStandard}
\alias{nakkeStandard}
\title{Server logic function for skiny Nakke}
\usage{
nakkeStandard(input, output, session, valgtVar, enhetsUtvalg, valgtMaal)
}
\arguments{
\item{input}{inngangsdata}

\item{output}{info som sendes ut}

\item{session}{Hmmm}

\item{valgtVar}{Variabelen det skal vises resultat for.
            Alder: Aldersfordeling
            AntallNivaaOpr: Antall nivå operert
            Antibiotika: Er det gitt antibiotikaprofylakse?
            Arbeidstaus12mnd: Arbeidsstatus 12 mnd. etter operasjon
            Arbeidstaus3mnd: Arbeidsstatus 3 mnd. etter operasjon
            ArbeidstausPreOp: Arbeidsstatus før operasjon
            ASAgrad: ASA-grad
            BMI: Pasientenes BMI (Body Mass Index)
            EqAngstPreOp: Helsetilstand, Angst
            ErstatningPreOp: Søkt erstatning?
            FornoydBeh12mnd: Fornøydhet med behandlinga på sykehuset, 12 mnd
            FornoydBeh3mnd: Fornøydhet med behandlinga på sykehuset, 3 mnd
            OperasjonsKategori: Hastegrad
            LiggeDognPostop: Antall liggedøgn postoperativt
            LiggeDognTotalt: Totalt antall liggedøgn
            Morsmal: Morsmål
            NytteOpr12mnd: Nytte av operasjon, 12 mnd
            NytteOpr3mnd: Nytte av operasjon, 3 mnd
            OprIndikPareseGrad: Paresegrad før operasjon
            Roker: Røyker pasienten?
            Saardren: Har pasienten fått sårdren?
            SivilStatus: Sivilstatus
            SmertestillBrukPreOp: Hyppighet av smertestillende før operasjonen
            Snuser: Snuser pasienten?
            SymptVarighetArmer: Varighet av utstrålende armsmerter
            SymptVarighetNakkeHode: Varighet av nakke-/hodesmerter
            TidlOpr: Er pasienten tidligere operert?
            TidlOprAntall: Antall tidligere operasjoner
            UforetrygdPreOp: Søkt uføretrygd?
            Utdanning: Utdanningsnivå

Detajer...:}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}
}
\description{
nakkeStandard is the module server function for shiny Nakke
}
