% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeVarTilrettelegg.R
\name{NakkeVarTilrettelegg}
\alias{NakkeVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NakkeVarTilrettelegg(RegData, valgtVar, ktr = 0, figurtype = "andeler")
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{parameter som angir hvilke(n) variabel man ønsker å tilrettelegge for videre
beregning.}

\item{ktr}{Angir om 3 eller 12 mnd kontroll. 0-ikke valgt (standard), 1-3mnd., 2-12mnd.,}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
}
