% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeRegDataSQL.R
\name{NGERHentRegData}
\alias{NGERHentRegData}
\title{Provide global dataframe for Nakke}
\usage{
NGERHentRegData(datoFra = "2012-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{- Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{- Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}
}
\value{
RegData data frame
}
\description{
Provides Nakke data from staging
}

