% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkePreprosesser.R
\name{NGERPreprosess}
\alias{NGERPreprosess}
\title{Preprosesser data fra Degenerativ Nakke}
\usage{
NGERPreprosess(RegData = RegData)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}
}
\value{
RegData En dataramme med det preprosesserte datasettet
}
\description{
Denne funksjonen definerer og formaterer variabler
}

