% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{PasMdblReg}
\alias{PasMdblReg}
\title{Finner pasienter med potensielt dobbeltregistrerte skjema}
\usage{
PasMdblReg(RegData, tidsavvik = 30)
}
\arguments{
\item{RegData}{dataramme fra nakkeregisteret}

\item{tidssavik}{- maks tidsavvik (dager) mellom to påfølgende registreringer som sjekkes}
}
\description{
Finner pasienter med potensielt dobbeltregistrerte skjema
}
