% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/runShinyAppReports.R
\name{runShinyAppReports}
\alias{runShinyAppReports}
\title{runShinyAppReports starts shiny reports}
\usage{
runShinyAppReports()
}
\description{
runShinyAppReports presents figs and stuff from the package in shiny
}

