% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{SorterOgNavngiTidsEnhet}
\alias{SorterOgNavngiTidsEnhet}
\title{Tilrettelegge tidsenhetvariabel. Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal}
\usage{
SorterOgNavngiTidsEnhet(RegData, tidsenhet = "Aar", tab = 0)
}
\arguments{
\item{RegData}{dataramme}

\item{tidsenhet}{valgt tidsenhet for visninga}

\item{tab}{Hmmm}
}
\description{
Tilrettelegge tidsenhetvariabel. Legger til tidsenhetene Aar, Halvaar, Mnd og Kvartal
}
