% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeTabeller.R
\name{tabAntOpphShAar}
\alias{tabAntOpphShAar}
\title{Antall opphold per år}
\usage{
tabAntOpphShAar(RegData, datoTil = Sys.Date(), antAar = 5)
}
\arguments{
\item{RegData}{dataramme med alle data}

\item{datoTil}{sluttdato}
}
\description{
Antall opphold per år
}
