% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeTabeller.R
\name{tabAntOpphShMnd}
\alias{tabAntOpphShMnd}
\title{Antall opphold siste X (antMnd) mnd}
\usage{
tabAntOpphShMnd(RegData, datoTil = Sys.Date(), antMnd = 6, reshID = 0)
}
\arguments{
\item{RegData}{dataramme med alle dato}

\item{datoTil}{sluttdato}

\item{antMnd}{antall måneder som skal vises}

\item{reshID}{reshID for filtrering}
}
\description{
Antall opphold siste X (antMnd) mnd
}
