% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeTabeller.R
\name{tabAntSkjema}
\alias{tabAntSkjema}
\title{Hvor mange skjema av hver type}
\usage{
tabAntSkjema(
  SkjemaOversikt,
  datoFra = "2019-01-01",
  datoTil = Sys.Date(),
  skjemastatus = 1
)
}
\arguments{
\item{SkjemaOversikt}{tabellen skjemaoversikt fra QReg}

\item{datoFra}{startdato}

\item{datoTil}{til og med dato}

\item{skjemastatus}{status på registreringsskjemaet}
}
\description{
Hvor mange skjema av hver type
}
