% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHjelpefunksjoner.R
\name{tilretteleggDataSKDE}
\alias{tilretteleggDataSKDE}
\title{Tilrettelegge data for offentlig visning.}
\usage{
tilretteleggDataSKDE(RegData = RegData, datoFra = "2014-01-01", aar = 0)
}
\arguments{
\item{RegData}{- data}

\item{datoFra}{- startdato}

\item{aar}{- velge hele år (flervalg)}

\item{valgtVar}{-}
}
\value{
Datafil til Resultatportalen
}
\description{
Tilrettelegge data for offentlig visning.
}
