% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/valgtVarFigAndeler.R
\name{valgtVarFigAndeler}
\alias{valgtVarFigAndeler}
\title{nakkeValgtVarFigAndeler provides a list of registry vars for the report}
\usage{
valgtVarFigAndeler()
}
\value{
valgtVarListFigAndeler list of names and values
}
\description{
The list of possible variables to be used in the report type
\emph{figAndeler}. Placed as its own function to trim the shiny app for
Nakke
}

