% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/valgtVarFigAndelerGrVar.R
\name{valgtVarFigAndelerGrVar}
\alias{valgtVarFigAndelerGrVar}
\title{nakkeValgtVarFigAndelerGrVar provides a list of registry vars for the report}
\usage{
valgtVarFigAndelerGrVar()
}
\value{
valgtVarListFigAndelerGrVar list of names and values
}
\description{
The list of possible variables to be used in the report type
\emph{figAndelerGrVar}. Placed as its own function to trim the shiny app for
Nakke
}

