% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/valgtVarFigAndelTid.R
\name{valgtVarFigAndelTid}
\alias{valgtVarFigAndelTid}
\title{valgtVarFigAndelTid provides a list of registry vars for the report}
\usage{
valgtVarFigAndelTid()
}
\value{
valgtVarListFigAndelTid list of names and values
}
\description{
The list of possible variables to be used in the report type
\emph{figAndelTid}. Placed as its own function to trim the shiny app for
Nakke
}

