% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeTabeller.R
\name{visVanligsteProcDiag}
\alias{visVanligsteProcDiag}
\title{Vise figurdata som tabell, sentralmål per sykshus}
\usage{
visVanligsteProcDiag(RegData, ant = 20, prosdiag = "pros")
}
\arguments{
\item{RegData}{dataramme}

\item{ant}{antall diagnoser/prosedyrer}

\item{prosdiag}{vise prosedyrer ('pros', standard) eller diagnoser ('diag')}
}
\description{
Vise figurdata som tabell, sentralmål per sykshus
}
