% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeHighcharts.R
\name{AndelerGrVarHighchart}
\alias{AndelerGrVarHighchart}
\title{Provider of Highchart object}
\usage{
AndelerGrVarHighchart(AndelerGrSort, Ngrtxt, AndelHele, N, Tittel, utvalgTxt,
  GrNavnSort, farger)
}
\value{
h1 Highchart object
}
\description{
Provide Highchart object for AndelerGrVar
}

