% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeHighcharts.R
\name{AndelerHighchart}
\alias{AndelerHighchart}
\title{Collection of functions to provide Highchart objects}
\usage{
AndelerHighchart(hoved, AntHoved, NHoved, rest, AntRest, NRest, retn, subtxt,
  grtxt, shtxt, fargeHoved, fargeRest, medSml, Tittel)
}
\value{
h1 Highchart object
}
\description{
Yes, still a collection of functions...
}

