% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Søylediagram, horisontalt eller vertikalt, som viser andeler av ulike variable:}
\usage{
FigAndeler(RegData, hentData = 1, valgtVar, datoFra = "2007-01-01",
  datoTil = "3000-12-31", minald = 0, maxald = 130, erMann = "",
  tittel = 1, outfile = "", reshID, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{- ei dataramme med alle nødvendige variable fra registeret}

\item{hentData}{- Settes til 1 (standard) om det skal spørres etter data fra funksjonen.
Settes til en annen verdi om data leveres til funksjonen gjennom 'RegData',
eksempelvis ved bruk av eksempeldatasettet eller ved kall fra andre
funksjoner der data allerede er tilgjengelig.}

\item{valgtVar}{- Variabelen det skal vises resultat for. Må velges: Alder, BMI, ...}

\item{datoFra}{- Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{- Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{- alder, fra og med}

\item{maxald}{- alder, til og med}

\item{erMann}{- kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{outfile}{- navn på fil figuren skrives ned til}

\item{enhetsUtvalg}{- 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{libkat}{- sti til bibliotekkatalog (fjerne..?)}

\item{egenReshID}{- avdelingsid (reshID) for egen avdeling,}
}
\description{
Her kommer muligens mer forklaring...?
}
\details{
Inndata:
}

