% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigAndelerGrVar.R
\name{FigAndelerGrVar}
\alias{FigAndelerGrVar}
\title{Søylediagram med andeler for hver grupperingsenhet (sykehus, RHF, ...)}
\usage{
FigAndelerGrVar(RegData, valgtVar, datoFra = "2012-01-01",
  datoTil = "3000-12-31", enhetsUtvalg = 0, minald = 0, maxald = 130,
  erMann = "", hentData = 0, preprosess = TRUE, tittel = 1, reshID,
  outfile = "")
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Variabelen det skal vises resultat for.
Alder: Aldersfordeling
AndreRelSykdommer: Andre sykdommer
Antibiotika: Fått antibiotika
Arbeidstaus12mnd: Mottar sykepenger, 12 mnd etter operasjon?
Arbeidstaus3mnd: Mottar sykepenger, 3 mnd etter operasjon?
ArbeidstausPreOp: Mottar sykepenger, preoperativt?
ASAgrad: ASA-grad > II
BMI: Pasienter med fedme
EnhverKompl3mnd: Alle komplikasjoner
ErstatningPreOp: Søkt/planlegger å søke erstatning
FornoydBeh12mnd: Fornøyde pasienter, 12 mnd.
FornoydBeh3mnd: Fornøyde pasienter, 3 mnd.
KomplinfekDyp3mnd: Pasientrapportert dyp infeksjon, 3 mnd.
KomplinfekOverfl3mnd: Overfladisk infeksjon, 3 mnd.
KomplStemme3mnd: Stemmevansker, 3 mnd.
KomplSvelging3mnd: Svelgvansker, 3 mnd.
Misfor12mnd: Misfornøyde pasienter, 12 mnd.
Misfor3mnd: Misfornøyde pasienter, 3 mnd.
NDIendr12mnd: Minst 30% forbedring av NDI, 12 mnd.
NRSsmerteArmEndr12mnd: Minst 30% forbedring av NSR-arm, 12 mnd.
NytteOpr12mnd: Klart bedre, 12 mnd.
NytteOpr3mnd: Klart bedre, 3 mnd.
Verre12mnd: Klart verre, 12 mnd.
Verre3mnd. Klart verre, 3 mnd.
OprIndikMyelopati: Operasjonsårsak, Myelopati
Roker: Røykere
Saardren: Andel som får sårdren
SmertestillPreOp: Bruker smertestillende, preop.
SymptVarighetNakkeHode: Varighet av hode-/nakkesmerter over 1 år
SymptVarighetSmerterUker: Varighet av smerter minst 1 år
UforetrygdPreOp: Søkt eller planlegger å søke uføretrygd?
Utdanning: Andel høyskole-/universitetsutdannede}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{tittel}{Om tittel skal vises i figuren eller ikke. Tittel tas bort i samlerapporter.
0: ikke vis tittel, 1: vis tittel (standard)}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}

\item{outfile}{Navn på fil figuren skrives ned til}
}
\value{
Figur med...
}
\description{
Funksjon som genererer en figur med andeler av en variabel for en valgt grupperingsvariabel,
f.eks. sykehus.
Funksjonen er delvis skrevet for å kunne brukes til andre grupperingsvariable enn sykehus
}

