% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeFigGjsnTid.R
\name{FigGjsnTid}
\alias{FigGjsnTid}
\title{Tidsutvikling av gjennomsnitt/median for en gitt variabel}
\usage{
FigGjsnTid(RegData, outfile, valgtVar, erMann = "", minald = 0,
  maxald = 130, tittel = 1, datoFra = "2007-01-01",
  datoTil = "3000-01-01", valgtMaal = "", enhetsUtvalg = 1,
  hentData = 0, preprosess = TRUE, reshID)
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{outfile}{Navn på fil figuren skrives ned til}

\item{valgtVar}{- Variabelen det skal vises resultat for.
EMSendr12mnd: Forbedring av EMS hos myelopati-pasienter, 12 mnd.
EMSendr3mnd: Forbedring av EMS hos myelopati-pasienter, 3 mnd.
EQ5Dendr12mnd: Forbedring av EQ5D, 12 mnd.
EQ5Dendr3mnd: Forbedring av EQ5D, 3 mnd.
Eq5DScorePreOp: EQ5D, før operasjon
KnivtidTotalMin: total knivtid
LiggeDognPostop: liggetid etter operasjon
LiggeDognTotalt: antall liggedøgn, totalt
NDIscorePreOp: NDI før operasjon
NDIendr3mnd: Forbedring av NDI, 3 mnd. etter operasjon
NDIendr12mnd: Forbedring av NDI, 3 mnd. etter operasjon}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{tittel}{Om tittel skal vises i figuren eller ikke. Tittel tas bort i samlerapporter.
0: ikke vis tittel, 1: vis tittel (standard)}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{valgtMaal}{Sentralmål: 'Med' gir median, alt annet gir gjennomsnitt}

\item{enhetsUtvalg}{Sammenlikning eller ikke: 0-hele landet, 1-egen enhet mot resten av landet, 2-egen enhet}

\item{hentData}{Angir om funksjonen skal kjøre spørring for å hente data eller ikke.
0: ikke kjør (standard)
1: kjør}

\item{preprosess}{Skal data preprosesseres, dvs. gjøre standard omregning av variable og beregne nye.
TRUE (standard) / FALSE}

\item{reshID}{Avdelingsid (reshID) for egen avdeling,}
}
\value{
Figur med...
}
\description{
Gjennomsnitt/median per år med konfidensintervall for valgt variabel.
Hvis man har valgt å sammenlikne, vises  konfidensintervall for resten av landet (evt. annen
sammenlikningsgruppe) i bakgrunnen.
}
\details{
Smerte: Skalaen går fra 0 til 10, dvs. at differansen ligger mellom -10 og 10..
EQ5D: Skala fra -0.594 tl 1, jo høyere jo bedre.
Oswestry: Skala fra 0 til 100, hvor lavest er friskest
}

