% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeHentData.R
\name{hentDataTabell}
\alias{hentDataTabell}
\title{Hent datatabell fra nakkes database}
\usage{
hentDataTabell(
  tabellnavn = "surgeonform",
  qVar = "*",
  datoFra = "2023-01-01",
  datoTil = Sys.Date(),
  egneVarNavn = 1
)
}
\arguments{
\item{tabellnavn}{Navn på tabell som skal lastes inn.}

\item{egneVarNavn}{0 - Qreg-navn benyttes.
                   1 - selvvalgte navn fra Friendlyvar benyttes
Bare ferdigstilte (status=1) legeskjema og pasientskjema overføres
(Sjekket 23.feb 2026)}
}
\description{
Hent datatabell fra nakkes database
}
