% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeLibUtvalg.R
\name{NakkeLibUtvalg}
\alias{NakkeLibUtvalg}
\title{Funksjon som gjør utvalg i datagrunnlaget til registreringene for Nakke}
\usage{
NakkeLibUtvalg(RegData, datoFra, datoTil, minald = 0, maxald = 130,
  erMann = "", fargepalett = "BlaaOff")
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}

\item{minald}{Alder, fra og med}

\item{maxald}{Alder, til og med}

\item{erMann}{Kjønn, 1-menn, 0-kvinner, standard: '' (alt annet enn 0 og 1), dvs. begge}

\item{fargepalett}{- Velge fargepalett, standard:BlaaOff ("offentliggjøringsfargene")}
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.
}

