% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NakkeRegDataSQL.R
\name{NakkeRegDataSQL}
\alias{NakkeRegDataSQL}
\title{Henter data registrert for Degenerativ Nakke}
\usage{
NakkeRegDataSQL(datoFra = "2012-01-01", datoTil = "2099-01-01")
}
\arguments{
\item{datoFra}{Operasjonsdato, fra og med. Standard: '2012-01-01'}

\item{datoTil}{Operasjonsdato, til og med. Standard: '3000-01-01' (siste registreringsdato)}
}
\value{
Henter dataramma RegData for Degenerativ Nakke
}
\description{
Henter data for Degenerativ Nakke fra "staging"
}

