% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/nakkeStandard.R
\name{nakkeStandard}
\alias{nakkeStandard}
\title{Server logic function for skiny Nakke}
\usage{
nakkeStandard(input, output, session, andelerValgtVar, enhetsUtvalg)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
nakkeStandard is the module server function for shiny Nakke
}

