% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakkeStandard.R
\name{nakkeStandard}
\alias{nakkeStandard}
\title{Server logic function for skiny Nakke}
\usage{
nakkeStandard(input, output, session, valgtVar, enhetsUtvalg, valgtMaal)
}
\description{
nakkeStandard is the module server function for shiny Nakke
}

