% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakkeStandardInput.R
\name{nakkeStandardInput}
\alias{nakkeStandardInput}
\title{Modulised UI for shiny}
\usage{
nakkeStandardInput(id, label = "Brukervalg")
}
\arguments{
\item{id}{namespace for this module}

\item{label}{Label for this module}
}
\description{
nakkeStandardInput provides common ui uc for shiny Nakke
}
