% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nakkeStandardInput.R
\name{nakkeStandardInput}
\alias{nakkeStandardInput}
\title{Modulised UI for shiny}
\usage{
nakkeStandardInput(id, label = "Brukervalg")
}
\description{
nakkeStandardInput provides common ui uc for shiny Nakke
}
\details{
@param id namespace for this module
 @param label Label for this module
 @export
}

