% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NakkeVarTilrettelegg.R
\name{NakkeVarTilrettelegg}
\alias{NakkeVarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NakkeVarTilrettelegg(RegData, valgtVar, ktr = 0, figurtype = "andeler")
}
\arguments{
\item{RegData}{Dataramme med alle nødvendige variable fra registeret}

\item{valgtVar}{Variabelen det skal vises resultat for.
            Alder: Aldersfordeling
            AntallNivaaOpr: Antall nivå operert
            Antibiotika: Er det gitt antibiotikaprofylakse?
            Arbeidstaus12mnd: Arbeidsstatus 12 mnd. etter operasjon
            Arbeidstaus3mnd: Arbeidsstatus 3 mnd. etter operasjon
            ArbeidstausPreOp: Arbeidsstatus før operasjon
            ASAgrad: ASA-grad
            BMI: Pasientenes BMI (Body Mass Index)
            EqAngstPreOp: Helsetilstand, Angst
            ErstatningPreOp: Søkt erstatning?
            FornoydBeh12mnd: Fornøydhet med behandlinga på sykehuset, 12 mnd
            FornoydBeh3mnd: Fornøydhet med behandlinga på sykehuset, 3 mnd
            OperasjonsKategori: Hastegrad
            LiggeDognPostop: Antall liggedøgn postoperativt
            LiggeDognTotalt: Totalt antall liggedøgn
            Morsmal: Morsmål
            NytteOpr12mnd: Nytte av operasjon, 12 mnd
            NytteOpr3mnd: Nytte av operasjon, 3 mnd
            OprIndikPareseGrad: Paresegrad før operasjon
            Roker: Røyker pasienten?
            Saardren: Har pasienten fått sårdren?
            SivilStatus: Sivilstatus
            SmertestillBrukPreOp: Hyppighet av smertestillende før operasjonen
            Snuser: Snuser pasienten?
            SymptVarighetArmer: Varighet av utstrålende armsmerter
            SymptVarighetNakkeHode: Varighet av nakke-/hodesmerter
            TidlOpr: Er pasienten tidligere operert?
            TidlOprAntall: Antall tidligere operasjoner
            UforetrygdPreOp: Søkt uføretrygd?
            Utdanning: Utdanningsnivå

Detajer...:}

\item{figurtype}{Hvilken figurtype det skal tilrettelegges variable for:
'andeler', 'andelGrVar', 'andelTid', 'gjsnGrVar', 'gjsnTid'}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
}
